Separate chart classes into files
