+++ Source/Extensions/Blazorise.Charts/BaseChart.cs
@@ -4,74 +4,14 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Blazorise.Utilities;
 using Microsoft.AspNetCore.Components;
-using Microsoft.JSInterop;
 #endregion
 
 namespace Blazorise.Charts
 {
-    /// <summary>
-    /// Interface is needed to set the value from javascript because calling generic component directly is not supported by Blazor.
-    /// </summary>
-    public interface IBaseChart
-    {
-        Task Event( string eventName, int datasetIndex, int index, string model );
-    }
-
-    public class BaseChart<TItem> : BaseComponent
-    {
-        #region Members
-
-        #endregion
-
-        #region Methods
-
-        protected override void BuildClasses( ClassBuilder builder )
-        {
-            builder.Append( ClassProvider.Chart() );
-
-            base.BuildClasses( builder );
-        }
-
-        protected override void Dispose( bool disposing )
-        {
-            if ( disposing )
-            {
-                _ = JS.Destroy( JSRuntime, ElementId );
-                JS.DisposeDotNetObjectRef( DotNetObjectRef );
-            }
-
-            base.Dispose( disposing );
-        }
-
-        #endregion
-
-        #region Properties
-
-        /// <inheritdoc/>
-        protected override bool ShouldAutoGenerateId => true;
-
-        protected DotNetObjectReference<ChartAdapter> DotNetObjectRef { get; set; }
-
-        [Inject] protected IJSRuntime JSRuntime { get; set; }
-
-        /// <summary>
-        /// Defines the chart data.
-        /// </summary>
-        [Parameter] public ChartData<TItem> Data { get; set; }
-
-        [Parameter] public EventCallback<ChartMouseEventArgs> Clicked { get; set; }
-
-        [Parameter] public EventCallback<ChartMouseEventArgs> Hovered { get; set; }
-
-        [Parameter] public RenderFragment ChildContent { get; set; }
-
-        #endregion
-    }
-
     public class BaseChart<TDataSet, TItem, TOptions, TModel> : BaseChart<TItem>, IBaseChart
-    where TDataSet : ChartDataset<TItem>
-    where TOptions : ChartOptions
-    where TModel : ChartModel
+        where TDataSet : ChartDataset<TItem>
+        where TOptions : ChartOptions
+        where TModel : ChartModel
     {
         #region Members
