+++ Source/Extensions/Blazorise.Charts/BaseChartOfT.cs
@@ -1,0 +1,59 @@
+﻿#region Using directives
+using Blazorise.Utilities;
+using Microsoft.AspNetCore.Components;
+using Microsoft.JSInterop;
+#endregion
+
+namespace Blazorise.Charts
+{
+    /// <summary>
+    /// Base class for all chart types.
+    /// </summary>
+    /// <typeparam name="TItem">Generic dataset value type.</typeparam>
+    public class BaseChart<TItem> : BaseComponent
+    {
+        #region Methods
+
+        protected override void BuildClasses( ClassBuilder builder )
+        {
+            builder.Append( ClassProvider.Chart() );
+
+            base.BuildClasses( builder );
+        }
+
+        protected override void Dispose( bool disposing )
+        {
+            if ( disposing )
+            {
+                _ = JS.Destroy( JSRuntime, ElementId );
+                JS.DisposeDotNetObjectRef( DotNetObjectRef );
+            }
+
+            base.Dispose( disposing );
+        }
+
+        #endregion
+
+        #region Properties
+
+        /// <inheritdoc/>
+        protected override bool ShouldAutoGenerateId => true;
+
+        protected DotNetObjectReference<ChartAdapter> DotNetObjectRef { get; set; }
+
+        [Inject] protected IJSRuntime JSRuntime { get; set; }
+
+        /// <summary>
+        /// Defines the chart data.
+        /// </summary>
+        [Parameter] public ChartData<TItem> Data { get; set; }
+
+        [Parameter] public EventCallback<ChartMouseEventArgs> Clicked { get; set; }
+
+        [Parameter] public EventCallback<ChartMouseEventArgs> Hovered { get; set; }
+
+        [Parameter] public RenderFragment ChildContent { get; set; }
+
+        #endregion
+    }
+}
